+++ lib/sns-helper/types/config.ts
@@ -1,5 +1,7 @@
 type TopicConfig = {
 	name: string
+	fifoTopic?: boolean
+	contentBasedDeduplication?: boolean
 }
 
 export type SNSConfig = {